# p2/openapi.yaml
openapi: 3.0.0
info:
  title: Team Task Board API
  description: 团队任务看板后端 RESTful API
  version: 1.0.0
servers:
  - url: http://localhost:3001/api
    description: 本地开发服务器

tags:
  - name: Health
    description: 健康检查
  - name: Tasks
    description: 任务管理

paths:
  /health:
    get:
      tags:
        - Health
      summary: 健康检查
      description: 返回服务状态和当前时间戳
      operationId: getHealth
      responses:
        '200':
          description: 服务正常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatus'

  /tasks:
    get:
      tags:
        - Tasks
      summary: 获取所有任务
      description: 获取任务列表，按创建时间倒序排列
      operationId: listTasks
      responses:
        '200':
          description: 成功返回任务列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      tags:
        - Tasks
      summary: 创建任务
      description: 创建一个新任务
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: 任务创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /tasks/{id}:
    get:
      tags:
        - Tasks
      summary: 获取单个任务
      description: 根据 ID 获取任务详情
      operationId: getTask
      parameters:
        - name: id
          in: path
          required: true
          description: 任务 ID
          schema:
            type: integer
      responses:
        '200':
          description: 成功返回任务
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: 任务不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Tasks
      summary: 更新任务
      description: 更新任务的全部字段
      operationId: updateTask
      parameters:
        - name: id
          in: path
          required: true
          description: 任务 ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 任务不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Tasks
      summary: 删除任务
      description: 根据 ID 删除任务
      operationId: deleteTask
      parameters:
        - name: id
          in: path
          required: true
          description: 任务 ID
          schema:
            type: integer
      responses:
        '204':
          description: 删除成功
        '404':
          description: 任务不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /tasks/{id}/status:
    patch:
      tags:
        - Tasks
      summary: 更新任务状态
      description: 仅更新任务的状态字段
      operationId: updateTaskStatus
      parameters:
        - name: id
          in: path
          required: true
          description: 任务 ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [todo, in-progress, done]
                  description: 任务状态
      responses:
        '200':
          description: 状态更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: 无效的状态值
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 任务不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    HealthStatus:
      type: object
      required:
        - status
        - timestamp
      properties:
        status:
          type: string
          enum: [ok]
          example: ok
        timestamp:
          type: string
          format: date-time
          example: '2024-01-15T10:30:00.000Z'

    Task:
      type: object
      required:
        - id
        - title
        - description
        - priority
        - status
        - assignee
        - dueDate
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: 完成项目需求文档
        description:
          type: string
          example: 编写详细的产品需求文档
        priority:
          type: string
          enum: [high, medium, low]
          example: high
        status:
          type: string
          enum: [todo, in-progress, done]
          example: todo
        assignee:
          type: string
          example: 张三
        dueDate:
          type: string
          format: date
          nullable: true
          example: '2024-02-01'
        createdAt:
          type: string
          format: date-time
          example: '2024-01-15T10:30:00.000Z'
        updatedAt:
          type: string
          format: date-time
          example: '2024-01-15T10:30:00.000Z'

    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          example: 完成项目需求文档
        description:
          type: string
          default: ''
          example: 编写详细的产品需求文档
        priority:
          type: string
          enum: [high, medium, low]
          default: medium
          example: high
        status:
          type: string
          enum: [todo, in-progress, done]
          default: todo
          example: todo
        assignee:
          type: string
          default: ''
          example: 张三
        dueDate:
          type: string
          format: date
          nullable: true
          example: '2024-02-01'

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          example: 完成项目需求文档
        description:
          type: string
          example: 编写详细的产品需求文档
        priority:
          type: string
          enum: [high, medium, low]
          example: medium
        status:
          type: string
          enum: [todo, in-progress, done]
          example: in-progress
        assignee:
          type: string
          example: 李四
        dueDate:
          type: string
          format: date
          nullable: true
          example: '2024-02-15'

    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Task not found